% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRanges.R
\name{AddAlerts}
\alias{AddAlerts}
\title{Internal function to add alerts to the results}
\usage{
AddAlerts(fullresults, bg)
}
\arguments{
\item{fullresults}{fullresults}

\item{bg}{bg}
}
\description{
Internal function to add alerts to the results
}
