% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRanges.R
\name{BgPolygon}
\alias{BgPolygon}
\title{Internal function -- gets polygon around distribution to crop predictors before modeling}
\usage{
BgPolygon(thinned_points, buffer.polygon = c(5, 10))
}
\arguments{
\item{thinned_points}{thinned_points}

\item{buffer.polygon}{buffer.polygon}
}
\description{
Internal function -- gets polygon around distribution to crop predictors before modeling
}
