% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_from_points.R
\name{ClimateFromPoints}
\alias{ClimateFromPoints}
\title{Extract climate information from species points.}
\usage{
ClimateFromPoints(points, species = "", lat = "lat", lon = "lon", res = 5)
}
\arguments{
\item{points}{A data.frame of three columns containing species and coordinates}

\item{species}{A character string indicating name of column with species names}

\item{lat}{A character string indicating name of column with latitudes}

\item{lon}{character string indicating name of column with longitudes}

\item{res}{A number (2.5, 5 or 10) indicating resolution of climatic layers from where the climate data is being extracted.}
}
\description{
Function to extract climate information from species points.
}
