% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRanges.R
\name{ColinearityTest}
\alias{ColinearityTest}
\title{Detecting collinearity in predictors}
\usage{
ColinearityTest(bg, predictors)
}
\arguments{
\item{bg}{bg}

\item{predictors}{predictors}
}
\description{
Detecting collinearity in predictors
}
