% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_from_points.R
\name{Frost}
\alias{Frost}
\title{Get proportion of cells in the range of a species that experience some frost and a lot of frost every year.}
\usage{
Frost(ranges, res = 2.5)
}
\arguments{
\item{res}{Resolution of the climatic layers to be used (a numeric: 2.5, 5 or 10)}

\item{points}{A list of ranges named as different species}
}
\value{
A data.frame with proportions of some frost and a lot of frost for each species
}
\description{
Get proportion of cells in the range of a species that experience some frost and a lot of frost every year.
}
