% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_from_points.R
\name{GetClimate_current}
\alias{GetClimate_current}
\title{Get summary statistics for four variables (BIO1, BIO4, BIO12, BIO15) based on a list of ranges
Those variable represent: annual temperature, temperature seasonality, annual precipitation and
precipitation seasonality}
\usage{
GetClimate_current(ranges, res = 2.5)
}
\arguments{
\item{res}{Resolution of the climatic layers to be used (a numeric: 2.5, 5 or 10)}

\item{points}{A list of ranges named as different species}
}
\value{
A data.frame with means and standard deviations for each climatic variable
}
\description{
Get summary statistics for four variables (BIO1, BIO4, BIO12, BIO15) based on a list of ranges
Those variable represent: annual temperature, temperature seasonality, annual precipitation and
precipitation seasonality
}
