% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_from_points.R
\name{GetClimate_heterogeneity}
\alias{GetClimate_heterogeneity}
\title{Get summary statistics for the spatial heterogeneity of four variables (BIO1, BIO4, BIO12, BIO15) based on a list of ranges
Those variable are: annual temperature, temperature seasonality, annual precipitation and
precipitation seasonality. Means and standard deviations are taken from a layer of slopes for each variable}
\usage{
GetClimate_heterogeneity(ranges, res = 2.5)
}
\arguments{
\item{res}{Resolution of the climatic layers to be used (a numeric: 2.5, 5 or 10)}

\item{points}{A list of ranges named as different species}
}
\value{
A data.frame with means and standard deviations for each climatic variable
}
\description{
Get summary statistics for the spatial heterogeneity of four variables (BIO1, BIO4, BIO12, BIO15) based on a list of ranges
Those variable are: annual temperature, temperature seasonality, annual precipitation and
precipitation seasonality. Means and standard deviations are taken from a layer of slopes for each variable
}
