% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_from_points.R
\name{GetClimate_instability}
\alias{GetClimate_instability}
\title{Get summary statistics for the temporal instability of four variables (BIO1, BIO4, BIO12, BIO15) based on a list of ranges
Those variable represent: annual temperature, temperature seasonality, annual precipitation and
precipitation seasonality. Means and standard deviations are taken from a layer that represent the
pairwise comparisons among five paleoclim layers.}
\usage{
GetClimate_instability(ranges, res = "2_5m")
}
\arguments{
\item{res}{Resolution of the climatic layers to be used (a character: "2_5m,"5m" or "10m")}

\item{points}{A list of ranges named as different species}
}
\value{
A data.frame with means and standard deviations for each climatic variable
}
\description{
Get summary statistics for the temporal instability of four variables (BIO1, BIO4, BIO12, BIO15) based on a list of ranges
Those variable represent: annual temperature, temperature seasonality, annual precipitation and
precipitation seasonality. Means and standard deviations are taken from a layer that represent the
pairwise comparisons among five paleoclim layers.
}
