% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_from_points.R
\name{GetClimateSummStats}
\alias{GetClimateSummStats}
\title{Get summary statistics for climatic data of a list of species based on their distribution points}
\usage{
GetClimateSummStats(
  points,
  species = "species",
  lat = "decimalLatitude",
  lon = "decimalLongitude",
  res = 2.5
)
}
\arguments{
\item{points}{A data.frame of three columns containing species and coordinates}

\item{species}{A character string indicating name of column with species names}

\item{lat}{A character string indicating name of column with latitudes}

\item{lon}{character string indicating name of column with longitudes}

\item{res}{A number (2.5, 5 or 10) indicating resolution of climatic layers from where the climate data is being extracted.}
}
\value{
A list with summaru statistics for the 19 WorldClim variables plut altitude
}
\description{
Get summary statistics for climatic data of a list of species based on their distribution points
}
