% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRanges.R
\name{GetRanges}
\alias{GetRanges}
\title{Estimates range and range size based on species distribution modeling.}
\usage{
GetRanges(
  points,
  species = "species",
  lat = "decimalLatitude",
  lon = "decimalLongitude",
  threshold = 0.75,
  buffer = 25,
  res = 10
)
}
\arguments{
\item{points}{A data.frame of distribution points with at least three columns where one column represents species names and the other two decimal coordinates}

\item{species}{The name of the column in the data.frame with the names of species}

\item{lat}{The name of the column in the data.frame with latitudes}

\item{lon}{The name of the column in the data.frame with longitudes}

\item{threshold}{A number between 0 and 1 for the threshold to make models binary}

\item{buffer}{The radius in kilometers around points to estimate range size when there are three or less valid points}

\item{res}{The resolution (2.5, 5 or 10) of climatic variables used for modeling}
}
\value{
A list containing information about the species distribution modeling performance and a raster with the possible range of each species
}
\description{
Estimates range and range size based on species distribution modeling.
}
