% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTraitDistribution.R
\name{GetResDistribution}
\alias{GetResDistribution}
\title{A function to map distribution of traits}
\usage{
GetResDistribution(raster1, raster2)
}
\arguments{
\item{raster1}{A raster of mapped trait distribution or species-richness}

\item{raster2}{A raster of mapped trait distribution or species-richness}
}
\value{
A raster with mapped residuals of a linear regression between raster1 and raster2
}
\description{
A function to map distribution of traits
}
