% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTraitDistribution.R
\name{GetSpRichness}
\alias{GetSpRichness}
\title{A function to overlay distributions and create a map of species richness}
\usage{
GetSpRichness(list_of_ranges)
}
\arguments{
\item{list_of_ranges}{A list in the format of the output of GetRanges()}
}
\value{
A raster of species richness
}
\description{
A function to overlay distributions and create a map of species richness
}
