% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTraitDistribution.R
\name{GetTraitDistribution}
\alias{GetTraitDistribution}
\title{A function to map distribution of traits (still with no PW)}
\usage{
GetTraitDistribution(
  list_of_ranges,
  trait_data,
  type = c("binary", "continuous")
)
}
\arguments{
\item{list_of_ranges}{A list in the format of the output of GetRanges()}

\item{trait_data}{A data.frame where the first column contain species names and the second contains trait data}

\item{type}{"binary" or "continuous", depending on the kind of trait data}
}
\value{
A raster with the traits mapped in space
}
\description{
A function to map distribution of traits (still with no PW)
}
