% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRanges.R
\name{LoadWcLayers}
\alias{LoadWcLayers}
\title{Internal function for now -- loads climatic layers from Worldclim}
\usage{
LoadWcLayers(res.layers)
}
\arguments{
\item{res.layers}{res.layers}
}
\description{
Internal function for now -- loads climatic layers from Worldclim
}
