% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRanges.R
\name{RangeFromFewPoints}
\alias{RangeFromFewPoints}
\title{Getting a proxy for range size when you have three points or less.
When there are three or less valid points, the distribution will be estimated from a circle around each point of radius equal to the argument buffer.}
\usage{
RangeFromFewPoints(thinned_points, predictors, buffer)
}
\arguments{
\item{thinned_points}{thinned_points}

\item{predictors}{predictors}

\item{buffer}{buffer}
}
\description{
Getting a proxy for range size when you have three points or less.
When there are three or less valid points, the distribution will be estimated from a circle around each point of radius equal to the argument buffer.
}
