% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRanges.R
\name{RangeSize}
\alias{RangeSize}
\title{Estimates range size from binary model}
\usage{
RangeSize(list_of_model_results, threshold)
}
\arguments{
\item{list_of_model_results}{list_of_model_results}

\item{threshold}{threshold}
}
\description{
Estimates range size from binary model
}
