% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullGbifCleaning.R
\name{RemoveCentroids}
\alias{RemoveCentroids}
\title{Removes points that are located in country centroids}
\usage{
RemoveCentroids(
  points,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  buffer = 75000
)
}
\arguments{
\item{points}{A data.frame of distribution points with at least three columns where one column represents species names and the other two decimal coordinates}

\item{lon}{The name of the column in the data.frame with longitudes}

\item{lat}{The name of the column in the data.frame with latitudes}

\item{buffer}{A number in meters around each country centroid for points to be removed}
}
\description{
Removes points that are located in country centroids
}
