% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullGbifCleaning.R
\name{RemoveCropsAndInvasive}
\alias{RemoveCropsAndInvasive}
\title{Removes invasive and crop species. Names of species must be given without author name and following the GBIF taxonomic backbone}
\usage{
RemoveCropsAndInvasive(points, species = "species")
}
\arguments{
\item{points}{A data.frame of distribution points with at least three columns where one column represents species names and other two decimal coordinates.}

\item{species}{The name of the column in the data.frame with the names of species}
}
\description{
Removes invasive and crop species. Names of species must be given without author name and following the GBIF taxonomic backbone
}
