% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullGbifCleaning.R
\name{RemoveDuplicates}
\alias{RemoveDuplicates}
\title{Removes duplicated species, latitudes and longitudes keeping only one of the duplicated values per species}
\usage{
RemoveDuplicates(
  points,
  species = "species",
  lon = "decimalLongitude",
  lat = "decimalLatitude"
)
}
\arguments{
\item{points}{A data.frame of distribution points with at least five columns where one column represents species names and other two decimal coordinates.}

\item{species}{The name of the column in the data.frame with the names of species}

\item{lon}{The name of the column in the data.frame with longitudes}

\item{lat}{The name of the column in the data.frame with latitudes}
}
\description{
Removes duplicated species, latitudes and longitudes keeping only one of the duplicated values per species
}
