% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullGbifCleaning.R
\name{RemoveHerbariaLocalities}
\alias{RemoveHerbariaLocalities}
\title{Removes points around large herbaria.}
\usage{
RemoveHerbariaLocalities(
  points,
  lon = "decimalLongitude",
  lat = "decimalLatitude"
)
}
\arguments{
\item{points}{A data.frame of distribution points with at least three columns where one column represents species names and other two decimal coordinates.}

\item{lon}{The name of the column in the data.frame with longitudes}

\item{lat}{The name of the column in the data.frame with latitudes}
}
\description{
Removes points around large herbaria.
}
