% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullGbifCleaning.R
\name{RemoveOutliers}
\alias{RemoveOutliers}
\title{Removes outliers}
\usage{
RemoveOutliers(
  points,
  species = "species",
  lon = "decimalLongitude",
  lat = "decimalLatitude"
)
}
\arguments{
\item{points}{A data.frame of distribution points with at least three columns where one column represents species names and the other two decimal coordinates}

\item{species}{The name of the column in the data.frame with the names of species}

\item{lon}{The name of the column in the data.frame with longitudes}

\item{lat}{The name of the column in the data.frame with latitudes}
}
\description{
Removes outliers
}
