% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullGbifCleaning.R
\name{RemoveSeaPoints}
\alias{RemoveSeaPoints}
\title{Removes points in the sea}
\usage{
RemoveSeaPoints(
  points,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  buffer = 0
)
}
\arguments{
\item{points}{A data.frame of distribution points with at least three columns where one column represents species names and the other two decimal coordinates}

\item{lon}{The name of the column in the data.frame with longitudes}

\item{lat}{The name of the column in the data.frame with latitudes}

\item{buffer}{A number of degrees around continental areas where points are still kept after cleaning}
}
\description{
Removes points in the sea
}
