% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullGbifCleaning.R
\name{RemoveWrongCountries}
\alias{RemoveWrongCountries}
\title{Removes points that are located in the wrong country according to their GBIF labels
That will remove points that are not located in the countries where their labels say they were
collected}
\usage{
RemoveWrongCountries(
  points,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  buffer = 5
)
}
\arguments{
\item{points}{A data.frame of distribution points with at least five columns where one column represents species names and other two decimal coordinates.}

\item{lon}{The name of the column in the data.frame with longitudes}

\item{lat}{The name of the column in the data.frame with latitudes}

\item{buffer}{A number of degrees around each country where points are still consider part of that country}
}
\description{
Removes points that are located in the wrong country according to their GBIF labels
That will remove points that are not located in the countries where their labels say they were
collected
}
\details{
For now, the input data.frame must have a column named countryCode and one named gbifID, as the .csv files downloaded directly from GBIF.
}
