% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullGbifCleaning.R
\name{RemoveZeros}
\alias{RemoveZeros}
\title{Removes points that have 0 for both latitude and longitude}
\usage{
RemoveZeros(points, lon = "decimalLongitude", lat = "decimalLatitude")
}
\arguments{
\item{points}{A data.frame of distribution points with at least three columns where one column represents species names and the other two decimal coordinates}

\item{lon}{The name of the column in the data.frame with longitudes}

\item{lat}{The name of the column in the data.frame with latitudes}
}
\description{
Removes points that have 0 for both latitude and longitude
}
