% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_from_points.R
\name{Thinning}
\alias{Thinning}
\title{Thinning distribution data to smooth sampling bias}
\usage{
Thinning(
  points,
  species = "species",
  lat = "decimalLatitude",
  lon = "decimalLongitude",
  n = 1
)
}
\arguments{
\item{points}{A data.frame of three columns containing species and coordinates}

\item{species}{A character string indicating name of column with species names}

\item{lat}{A character string indicating name of column with latitudes}

\item{lon}{character string indicating name of column with longitudes}

\item{n}{A number indicating how many points to keep in each cell after thinning}
}
\description{
Thinning distribution data to smooth sampling bias
}
