% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRanges.R
\name{Thinning_internal}
\alias{Thinning_internal}
\title{Performs spatial thinning before modeling by selecting one occurence per grid cell}
\usage{
Thinning_internal(points_for_range)
}
\arguments{
\item{points_for_range}{points_for_range}
}
\description{
Performs spatial thinning before modeling by selecting one occurence per grid cell
}
