% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullGbifCleaning.R
\name{UnusualDistributions}
\alias{UnusualDistributions}
\title{Flag species that have weird distributions in the dataset}
\usage{
UnusualDistributions(
  points,
  species = "species",
  lat = "decimalLatitude",
  lon = "decimalLongitude",
  buffer.polygon = c(5, 10)
)
}
\arguments{
\item{points}{A data.frame of distribution points with at least three columns where one column represents species names and other two decimal coordinates.}

\item{lat}{The name of the column in the data.frame with latitudes}

\item{lon}{The name of the column in the data.frame with longitudes}
}
\description{
Flag species that have weird distributions in the dataset
}
